Some performance tweaks for testing.
Assert.* methods take a `msg` argument, which is preformatted.  This
means that we end up taking the hit of formatting arrays/etc even
if the assert passes.  Change to formatting internally to avoid
this.

Formatting the expression stack seems to take longer than we'd like.
Capturing just the frame doesn't work for some reason, so make
it an optional flag.

Try to pretty print stacks a little bit.  Common regions of the
stacks should be removed as well at some point.