Refactor concatenate mapper and extent translation
This fixes the negative upper left boundary bug in translate extent that
occurs when tmp_ul is negative. The negative dimension in the upper left
coordinate will instead be set to 0.

The column logic and 1d if branch are replaced with a tight if, elif,
else branch that sets the row in result to either data_a, data_b, or
data_a + data_b (using np.concatenate).