Fix tests timing and return value in mapper.
Glom the assign expression before the assert statement, otherwise
from_numpy gets executed after the ndarray has already been changed
(rendering the test useless). To fix this, the tests glom the
RegionMapExpr first, before the corresponding ndarray has been updated.

Because of this bug, the assign mapper bug wasn't being detected. The
assign mapper should return value, not tile. Assign mapper was kept in a
separate function (instead of a lambda expression) so debugging would be
made easier (and in case a better implementation of the mapper is
found).