Increase NUM_BODIES to 1000; add optimization flag.
On beaker-22, this takes ~38 seconds. 50 bodies take ~36 seconds, so as
the number of bodies increases from 50 to 1000, the time does not
increase significantly. (4 workers). As the number of bodies increases,
the variance in time per step does increase (though the average hovers
around 1.80 second per move, for both 50 and 1000).

However, above 2000, the computer runs out of memory and is killed by
the OS. Even increasing the number of workers to 32 does not stop the
machine from running out of memory.