Reduce memory usage; fix nbody answer.
We reduce memory usage by setting `target` in set_diagonal (so shape
is known and does not have to be calculated).

This commit fixes set_diagonal when tile not on diagonal and removes
dead code.

This removes unneccessary transpose calls that can be replaced with
a smarter reshape. Default timestep in benchmark changed to 50.