Fix #184; fix cluster mode for interactive repls.
This commit removes the `FileWatchdog` class. All of the ssh connections
used to initialize spartan are closed after use. These were responsible
for sluggish user interaction with a repl.

To shutdown spartan in cluster mode, this commit installs a shutdown
handler on `SIGINT` (`ctrl-C`), which calls `spartan.shutdown()` and
uninstalls the signal.